Rename histogram module
